+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-* Using spyglass that's held in offhand is now supported
-* Fixed broken zooming animation of your character in multiplayer
-* When using a fake spyglass (in creative/spectator), a spyglass sound will be played
+* Updated to 1.21.8